Move review questions service to an independent module
